regionPlot

Function to plot a particular region from a single chromosomes for a sample

regionPlot(cnv_file, from = NA, to = NA, chrom = "X", ylim = c(-5, 5),
  tick = 100000, bp1 = NA, bp2 = NA, title = NA)

Arguments

cnv_file
File to plot. [Required]

from
Start coordinate of plotting window [Default from=2950000 ]

to
End coordinate of plotting window [Default to=3400000]

chrom
Specify the chromosome to plot [Default 'X']

ylim
Adjust y axis limits in plot [Default ylim=c(-5,5)]

tick
Adjust spacing of ticks [Defualt 1Mb]

bp1
Draw a vertical line at given coordinate to mark bp1

bp2
Draw a vertical line at given coordinate to mark bp2

title
Plot title

Examples

regionPlot(cnv_file="data/w500/test.window-500.cnv", from=3050000, to=3450000, chrom="X", ylim=c(-7,7), bp1=3129368,bp2=3352041, tick=100000, title="222Kb DEL on X")

#> Processing data/w500/test.window-500.cnv 
#> Specified region 3050000 - 3450000 on X 
#> Plotting ticks on X every 1e+05 
#> Drawing lines for breakpoints: bp1= 3129368  bp2= 3352041 
#> Chrom: X 
#> Specified ylim -7 7 
#> Plotting region X:3050000-3450000 for test

#> Warning: cannot open file 'data/w500/test.window-500.cnv': No such file or directory

#> Error in file(file, "rt"): cannot open the connection